fix platform related postgres timestamp precision issues in tests #2
